refactor: reorganize admin settings components and remove deprecated files
