+++ frontend/app/(admin)/admin/conversation/page.tsx
@@ -1,5 +1,5 @@
 import { AdminShell } from "@/features/admin/components/admin-shell";
-import { AdminConversationSettingsPage } from "@/features/admin/components/sections/settings/admin-settings-conversation";
+import { AdminConversationSettingsPage } from "@/features/admin/components/sections/conversation/admin-conversation";
 
 export default function AdminConversationSettingsRoute() {
   return (
